backup_code/requirements.txt
# Core Framework
google-adk==0.5.0
google-genai==1.14.0
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
deprecated==1.3.1
wrapt==2.0.1

# PDF Processing
PyMuPDF==1.25.5
pdf2image==1.17.0
Pillow==11.1.0

# Image Processing & AI Vision
google-cloud-vision==3.9.0
opencv-python==4.11.0.86

# Audio Processing
soundfile==0.13.1
numpy==2.2.5

# Web & Async
websockets==14.2
aiofiles==24.1.0
httpx==0.28.1

# Utilities
pydantic==2.10.6
typing-extensions==4.12.2

# OCR (for text extraction from comic panels)
pytesseract==0.3.13

# Session Management
cachetools==5.5.2
